gradio_dashboard/requirements.txt
gradio==3.29.0
google-genai==0.1.0
